# openapi.yaml
...
openapi: "3.0.3"
info:
  title: "identuum-idp-oss API"
  version: "identuum-idp-oss-api-docgen 0.2.0-p3"
  description: "Machine-generated OpenAPI 3 specification for the identuum-idp-oss API surface. Generated by tools/api-docgen from // docgen: source annotations. Response bodies without a known schema record are described but not field-modeled; see the referenced Go symbol in identuum-idp-oss for the authoritative shape."
paths:
...
  "/api/v1/organizations/{id}/roles/{role_id}/scopes":
    post:
      operationId: "identuum-idp-oss.rbac.add_role_scope"
      summary: "Add a scope to an org role."
      tags:
        - "rbac"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
        - name: "role_id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "Add a scope to an org role."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
  "/api/v1/organizations/{id}/roles/{role_id}/scopes/{scope_name}":
    delete:
      operationId: "identuum-idp-oss.rbac.remove_role_scope"
      summary: "Remove a scope from an org role."
      tags:
        - "rbac"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
        - name: "role_id"
          in: path
          required: true
          schema:
            type: string
        - name: "scope_name"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "Remove a scope from an org role."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
...
  "/api/v1/scope-templates":
    get:
      operationId: "identuum-idp-oss.scope-templates.list_scope_templates"
      summary: "List scope templates for the authenticated principal's organization (Authorization Server feature)."
      tags:
        - "scope-templates"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "List scope templates for the authenticated principal's organization (Authorization Server feature)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/oss_handlers_safeScopeTemplate"
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
    post:
      operationId: "identuum-idp-oss.scope-templates.create_scope_template"
      summary: "Create a scope template (Authorization Server feature)."
      tags:
        - "scope-templates"
      security:
        - cookieAuth: []
      responses:
        "201":
          description: "Create a scope template (Authorization Server feature)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/oss_handlers_safeScopeTemplate"
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
  "/api/v1/scope-templates/{id}":
    get:
      operationId: "identuum-idp-oss.scope-templates.get_scope_template"
      summary: "Show a single scope template by id (Authorization Server feature)."
      tags:
        - "scope-templates"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "Show a single scope template by id (Authorization Server feature)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/oss_handlers_safeScopeTemplate"
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
    put:
      operationId: "identuum-idp-oss.scope-templates.update_scope_template"
      summary: "Update a scope template (Authorization Server feature)."
      tags:
        - "scope-templates"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "Update a scope template (Authorization Server feature)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/oss_handlers_safeScopeTemplate"
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
    delete:
      operationId: "identuum-idp-oss.scope-templates.delete_scope_template"
      summary: "Delete a scope template (Authorization Server feature)."
      tags:
        - "scope-templates"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "Delete a scope template (Authorization Server feature)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: "identuum_session"
      description: "Browser session cookie backing a cookie-authenticated Principal (site_admin, org_admin, or any authenticated user, depending on the operation's actual role requirement — see its summary)."
    oauthClientAuth:
      type: http
      scheme: basic
      description: "OAuth 2.1 client authentication (client credentials / RFC 7592 management token) at client-management and token endpoints."
  schemas:
...
    oss_handlers_safeScopeTemplate:
      type: object
      description: "Response shape defined by the OSS handler package (schema record: oss.handlers.safeScopeTemplate). Field-level shape is not modeled here — see the identuum-idp-oss Go source for the authoritative struct."
...